Fix duplicate check.
In changeset 956e5f4b3a21 we tried to optimize the check for duplicate records
but we failed because what we need to check is that periods don't overlap
instead of checking that they're identical.

This commit reverts that patch while keeping the exit situation in which if the
user rejects and approves again the record, no new historical record of amounts
is created.